rename AdminPage to AdminNav
